Renamed: kiwixbuild/configs/{ios -> apple}.py
