+++ kiwixbuild/configs/__init__.py
@@ -1,3 +1,3 @@
 from .base import *
 
-from . import android, armhf, musl, flatpak, i586, ios, native, neutral, wasm
+from . import android, armhf, musl, flatpak, i586, apple, native, neutral, wasm
